add persistentlogin wrapper and finsish full working authentication and authorization
