+++ src/hooks/useAxiosPrivate.ts
@@ -22,7 +22,7 @@ export const useAxiosPrivate = () => {
       (response) => response,
       async (error) => {
         const prevRequest = error?.config;
-        if (error?.response?.staus === 401 && !prevRequest?.sent) {
+        if (error?.response?.status === 401 && !prevRequest?.sent) {
           prevRequest.sent = true;
           const newAccessToken = await refresh();
           prevRequest.headers['Authorization'] = `Bearer ${newAccessToken}`;
@@ -36,7 +36,7 @@ export const useAxiosPrivate = () => {
       axiosPrivate.interceptors.response.eject(responseIntercept);
       axiosPrivate.interceptors.request.eject(requestIntercept);
     };
-  }, [refresh, auth]);
+  }, [refresh, auth.accessToken]);
 
   return axiosPrivate;
 };
